Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main(){
     int a,b,c;
-    float avg;
+    double avg;
     scanf("%d %d %d",&a,&b,&c);
     sum = a+b+c;
-    avg= sum/3;
+    avg= sum/3.0;
     printf("Average: %.2f",avg);
     return 0;
 }
